update 10-22-22
aded the jumps, changed the values of the camera to follow player, changes the jumpForce so that a player can jump into a platform. Added more platforms to test the jumping. Flipped the sprite when running left or right. Added a bg sound and a jumping VFX but i might change the sound later.